refactoring

+++ src/main/java/com/tsystems/javaschool/ecare/dao/UserDAO.java
@@ -12,15 +12,20 @@ public class UserDAO implements IAbstractDAO<User>
 {
     private static volatile UserDAO instance;
 
-    private UserDAO() {
+    private UserDAO()
+    {
     }
 
-    public static UserDAO getInstance() {
+    public static UserDAO getInstance()
+    {
         UserDAO localInstance = instance;
-        if (localInstance == null) {
-            synchronized (UserDAO.class) {
+        if (localInstance == null)
+        {
+            synchronized (UserDAO.class)
+            {
                 localInstance = instance;
-                if (localInstance == null) {
+                if (localInstance == null)
+                {
                     instance = localInstance = new UserDAO();
                 }
             }
@@ -29,49 +34,58 @@ public class UserDAO implements IAbstractDAO<User>
     }
 
     @Override
-    public User saveOrUpdate(User cl) {
+    public User saveOrUpdate(User cl)
+    {
         return EntityManagerUtil.getEntityManager().merge(cl);
     }
 
     @Override
-    public User load(int id) {
+    public User load(int id)
+    {
         return EntityManagerUtil.getEntityManager().find(User.class, id);
     }
 
-    public User findUserByLoginAndPassword(String login, String password) {
+    public User findUserByLoginAndPassword(String login, String password)
+    {
         Query query = EntityManagerUtil.getEntityManager().createNamedQuery("User.findUserByLoginAndPassword", User.class);
         query.setParameter("login", login);
         query.setParameter("password", password);
         return (User) query.getSingleResult();
     }
 
-    public User findUserByNumber(long number) {
+    public User findUserByNumber(int number)
+    {
         Query query = EntityManagerUtil.getEntityManager().createNamedQuery("User.findUserByPhoneNumber", User.class);
         query.setParameter("number", number);
         return (User) query.getSingleResult();
     }
 
     @Override
-    public void delete(User cl) {
+    public void delete(User cl)
+    {
         EntityManagerUtil.getEntityManager().remove(cl);
     }
 
     @Override
-    public List<User> getAll() {
+    public List<User> getAll()
+    {
         return EntityManagerUtil.getEntityManager().createNamedQuery("User.getAllUsers", User.class).getResultList();
     }
 
     @Override
-    public void deleteAll() {
+    public void deleteAll()
+    {
         EntityManagerUtil.getEntityManager().createNamedQuery("Client.deleteAllClients").executeUpdate();
     }
 
     @Override
-    public long getCount() {
-        return ((Number)EntityManagerUtil.getEntityManager().createNamedQuery("Client.size").getSingleResult()).longValue();
+    public long getCount()
+    {
+        return ((Number) EntityManagerUtil.getEntityManager().createNamedQuery("Client.size").getSingleResult()).longValue();
     }
 
-    public User findUserByLogin(String login) {
+    public User findUserByLogin(String login)
+    {
         Query query = EntityManagerUtil.getEntityManager().createNamedQuery("Client.findClientByLogin", User.class);
         query.setParameter("login", login);
         return (User) query.getSingleResult();
